refactor(runtime): extract backpressure module (#2832)
